[RFC] pci: tegra: use new pci_register_host interface
Tegra is one of the remaining platforms that still use the
traditional pci_common_init_dev() interface for probing PCI
host bridges.

This demonstrates how to convert it to the pci_register_host
interface I just added in a previous patch. This leads to
a more linear probe sequence that can handle errors better
because we avoid callbacks into the driver, and it makes
the driver architecture independent.

As a side note, I should mention that I noticed this driver
does not register any IORESOURCE_IO resource with the bus,
but instead registers the I/O port window as a memory
resource, which is surely a bug.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>