[SUBMITTED 20160519] usb: chipidea: msm: remove open-coded phy init
After commit 233c7daf4eec ("usb: chipidea: msm: Initialize PHY on reset
event"), we always call usb_phy_init() when receiving a
CI_HDRC_CONTROLLER_RESET_EVENT event, but we also set the
USB_AHBBURST and USB_AHBMODE registers to the same values that
the reset function does.

This removes the duplicate initialization, to avoid the dependency
on the header file and the phy registers.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>