[SUBMITTED 20160617] lkdtm: avoid link error on ARM
The lkdtm_rodata file is built so that a function in it gets linked into
the .rodata ELF section. This works fine normally, but on 32-bit ARM
with really large kernels, it prevents the linker from inserting a
veneer for the call to __gnu_mcount_nc in case we are building with
"gcc -pg":

drivers/misc/built-in.o: In function `lkdtm_rodata_do_nothing':
panel.c:(.rodata+0x1480): relocation truncated to fit: R_ARM_CALL against symbol `__gnu_mcount_nc' defined in .text section in arch/arm/kernel/built-in.o

We probably never want the profiling code to be enabled for this file
anyway, so this just removes the option here, and for the sake of
consistency also in the lkdtm core module.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Fixes: 9a49a528dcf3 ("lkdtm: add function for testing .rodata section")