[SUBMITTED 20160620] [media] omap3isp: support 64-bit version of omap3isp_stat_data
C libraries with 64-bit time_t use an incompatible format for
struct omap3isp_stat_data. This changes the kernel code to
support either version, by moving over the normal handling
to the 64-bit variant, and adding compatiblity code to handle
the old binary format with the existing ioctl command code.

Fortunately, the command code includes the size of the structure,
so the difference gets handled automatically.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
---
This patch was originally part of a longer series, along with other patches
that we still have to rewrite for the generic ioctl interface, but the
OMAP specific ioctl commands are completely independent of that, so I think
we can just merge this version for 4.8.

in v4, I'm fixing a build regression that was introduced v3, which
separated this patch from the v4l_timeval type that I had in the
earlier versions.