y2038: convert ntp_validate_timex to use timespec64
As the type for timex->time is changing, the timeval_inject_offset_valid
check no longer works and we have to modify the caller anyway.

This version introduces a new timespec64_inject_offset_valid()
helper and uses that instead of the existing functions for
timespec and timeval that we want to get rid of anyway.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>