[SUBMITTED 20160704] [RFC] Kbuild: avoid "make tinyconfig" warnings
The introduction of "make *.config" as a shorthand for merging configuration
files unfortunately introduced some build warnings that we see in every
single run of the various build bots testing tinyconfig:

.config:966:warning: override: NOHIGHMEM changes choice state
.config:965:warning: override: SLOB changes choice state
.config:963:warning: override: KERNEL_XZ changes choice state
.config:962:warning: override: CC_OPTIMIZE_FOR_SIZE changes choice state
.config:933:warning: override: SLOB changes choice state
.config:930:warning: override: CC_OPTIMIZE_FOR_SIZE changes choice state
.config:870:warning: override: SLOB changes choice state
.config:868:warning: override: KERNEL_XZ changes choice state
.config:867:warning: override: CC_OPTIMIZE_FOR_SIZE changes choice state

This changes the behavior for tinyconfig again, so we call the special
allnoconfig target directly, with the fragment as the KCONFIG_ALLCONFIG
file so it gets merged into the configuration. The disadvantage of
this approach is that we cannot pass another KCONFIG_ALLCONFIG on
the command line, that would again have to be done using the mergeconfig
script afterwards.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Fixes: 63a91033d52e ("kbuild: add generic mergeconfig target, %.config")
Link: https://storage.kernelci.org/mainline/v4.7-rc6/x86-tinyconfig/build.log