[SUBMITTED 20160714] ARM: nuc900: make most header files local
Once we enable CONFIG_ARCH_MULTIPLATFORM, we no longer reference the include/mach/
directory for header files, so they become invisible to the existing users.

Since all users of the headers are now in arch/arm/mach-w90x900/, we can simply
move the headers to the same place and include them using "" instead of <>.

The only exception is mach/uncompress.h, which gets included by the zImage
uncompress code for non-multiplatform configurations, so the header remains
there for the moment, and I add the few macros it actually needs there to
let it be included standalone.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>