ARM: imx: remove bogus imx53_suspend_sz declaration
When CONFIG_SUSPEND is disabled, building the i.MX code results
in a harmless warning with "make W=1":

arch/arm/mach-imx/common.h:127:18: warning: 'imx53_suspend_sz' defined but not used [-Wunused-const-variable=]

As this warning is duplicated for each file that includes
common.h, it's worth looking into avoiding it. Fortunately,
the imx53_suspend/imx53_suspend_sz/imx6_suspend dummy
implementations are all unnecessary here and we can simply
make the caller conditional to avoid having references to
the symbols in object code.

The i.MX6 implementation already does this, and this patch
uses the same approach for i.MX53, which avoids the warnings
and reduces the size of the object code for the !SUSPEND
case.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>