crypto: aesni: avoid -Wmaybe-uninitialized warning, better
An earlier patch has hidden this warning:

arch/x86/crypto/aesni-intel_glue.c: In function ‘helper_rfc4106_decrypt’:
include/linux/scatterlist.h:67:31: warning: ‘dst_sg_walk.sg’ may be used uninitialized in this function [-Wmaybe-uninitialized]

but this is not a good workaround. To improve the code while addressing
the warning better, this reorganizes the code to avoid that variable and
have the shared code in a separate function to avoid some of the conditional
branches.

The resulting functions are a bit longer but also slightly less complex,
leaving no room for speculation on the part of the compiler.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>