afs: use random UUID
AFS uses a time based UUID to identify the host itself. This requires
getting a timestamp which is currently done through the getnstimeofday()
interface that we want to eventually get rid of.

Instead of replacing it with a ktime based interface, this simply
removes the entire function and uses generate_random_uuid() instead,
which has a v4 ("completely random") uuid instead of the time based one.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>