Input: cyapa - reduce stack size with KASAN
cyapa_gen5_show_baseline() has lots of local variables that are passed into
other functions by reference, and KASAN adds stack checks for each of them,
leading to rather large stack size that we can get a compiler warning for:

drivers/input/mouse/cyapa_gen5.c:2434:1: error: the frame size of 2032 bytes is larger than 1536 bytes [-Werror=frame-larger-than=]

I've tried to come up with a way to reduce this amount of stack size, and
my best idea was to consolidate most of them into arrays of three, as they
all come in max/min/avg tuples anyway. This reduces the stack size to around
800 bytes in the same configuration.

Alternatively we could possibly use a structure rather than an array, which
would be a little more code but possibly easier to read. I'm not entirely
happy with either approach, so hoping for other suggestions.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>