fat: reduce stack frame size for 64k pages
When updating a directory or the file allocation table, we do at most one
MMU page worth of writes at a time, and have a local array of pointers
to buffer heads. When the page size is 64K or larger, this can lead to
rather large kernel stacks, as seen from this warning:

fs/fat/fatent.c: In function 'fat_alloc_clusters':
fs/fat/fatent.c:548:1: error: the frame size of 1680 bytes is larger than 1536 bytes [-Werror=frame-larger-than=]

fat_alloc_clusters() is only ever asked to allocate one or two entries,
which requires no more than two buffer heads, so this is clearly too much.

The other three functions that have an array of buffer head pointers on
the stack use more: fat_alloc_new_dir and fat_add_new_entries will write
a full cluster and fat_free_clusters will update all entries that have
become unused, possibly the entire FAT, in page size units, so they do
potentially benefit from larger I/Os.

The page size limit seems arbitrary here, so I'm limiting the maximum
size to 16KB now for larger pages. This means that a kernel with 64K
pages now may have to do more I/Os than before but is still better off
than a kernel with 4K pages, and is less likely to cause a stack overflow
than before.

It might be worth optimizing this again to allow larger I/Os for
file deletion and directory creation, but if we do that, then it should
be done in a way that also helps on systems with 4K pages.

I have only build tested this patch, and am not too familiar with the
fatfs code, so it would be good to have a confirmation of my analysis
from someone who understands it better, or to get some testing on
a kernel with 64K pages.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>