caif: reduce stack size with KASAN
When CONFIG_KASAN is set, we can use relatively large amounts of kernel
stack space:

net/caif/cfctrl.c:555:1: warning: the frame size of 1600 bytes is larger than 1280 bytes [-Wframe-larger-than=]

This adds convenience wrappers around cfpkt_extr_head(), which is responsible
for most of the stack growth. When KASAN is enabled, we force the wrappers to
be out-of-line and also simplifies the endianess handling.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>