staging: rts5208: mark helpers as noinline_for_stack
rtsx_scsi_handler() and related functions contain of nested switch() statements
calling into all the other functions in this file. With KASAN enabled, this results
in rather high stack usage when the functions are inlined, as all local variables
of the leaf functions are put on the stack of the caller:

drivers/staging/rts5208/rtsx_scsi.c: In function 'app_cmd':
drivers/staging/rts5208/rtsx_scsi.c:2605:1: error: the frame size of 1544 bytes is larger than 1536 bytes [-Werror=frame-larger-than=]

This marks all of the functions as noinline_for_stack, which solves the problem
by needing only the stack space for the function we end up calling.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>