[SUBMITTED 20171019] time: remove timespec64 hack
This may be a somewhat controversial change, changing 64-bit architectures
to use the same 'struct timespec64' definition that 32-bit architectures
have, and removing a micro-optimization that tries to minimize the
difference between timespec and timespec64.

Starting with gcc-5, the compiler can completely optimize away the
timespec_to_timespec64 and timespec64_to_timespec functions on 64-bit
architectures. With older compilers, we introduce a couple of extra
copies of local variables, but those are easily avoided by using
the timespec64 based interfaces consistently, as we do in most of the
important code paths already.

The main upside of removing the hack is that printing the tv_sec
field of a timespec64 structure can now use the %lld format
string on all architectures without a cast to time64_t. Without
this patch, the field is a 'long' type and would have to be printed
using %ld on 64-bit architectures.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>